Add binary optimization and layouting tool (BOLT) (#2724)
Script to build libHyrise using PGO (profile-guided optimization) and BOLT (Binary Optimization and Layouting Tool) optimizations.

---------

+++ scripts/benchmark_all.sh
@@ -95,9 +95,22 @@ do
   # Checkout and build from scratch, tracking the compile time.
   git checkout "$commit"
   git submodule update --init --recursive
-  echo "Building $commit..."
-  $build_system clean
-  /usr/bin/time -p sh -c "( $build_system -j $(nproc) ${benchmarks} 2>&1 ) | tee benchmark_all_results/build_${commit}.log" 2>"benchmark_all_results/build_time_${commit}.txt"
+
+  # If there is a BOLT or PGO profile, we use it for building.
+  if [ -f ../resources/bolt.fdata ] || [ -f ../resources/libhyrise.profdata ]
+  then
+    echo "Detected bolt.fdata or libhyrise.profdata file, building $commit with PGO/BOLT..."
+    args="--num-cores $(nproc) --import-profile --build-system $build_system"
+    # Append `--no-bolt` or `--no-pgo` to the arguments if the respective files have not been found. The script then
+    # does not try to optimize using non-existant files.
+    [ ! -f ../resources/bolt.fdata ] && args="$args --no-bolt"
+    [ ! -f ../resources/libhyrise.profdata ] && args="$args --no-pgo"
+    /usr/bin/time -p sh -c "( python3 ../scripts/build_pgo_bolt.py $args 2>&1 ) | tee benchmark_all_results/build_${commit}.log" 2>"benchmark_all_results/build_time_${commit}.txt"
+  else
+    echo "Building $commit..."
+    $build_system clean
+    /usr/bin/time -p sh -c "( $build_system -j $(nproc) ${benchmarks} 2>&1 ) | tee benchmark_all_results/build_${commit}.log" 2>"benchmark_all_results/build_time_${commit}.txt"
+  fi
 
   # Run the benchmarks.
   cd ..  # hyriseBenchmarkJoinOrder needs to run from project root.
